Add extensions definition to gemspec (#455)
This fixes the issue that the AppSignal gem doesn't install through
Bundler when specifying a git source:

```
gem "appsignal",
  :git => "https://github.com/appsignal/appsignal-ruby.git",
  :branch => "master"
```

This change adds the `extensions` config option as described here:
https://guides.rubygems.org/specification-reference/#extensions
by default for the Ruby C-extension. This means it does not work for
JRuby when installing through a git source.

This was broken in #377 where we generate the gemspecs for Ruby and
JRuby during the "rake publish" process.

We only ask users to use the git source when testing something small. So
full support for all implementations (JRuby) doesn't seem necessary at
this time. We don't recommend users to rely on the git source
installation method all the time.