Fix DelayedJob name exctraction
In our documentation ( https://docs.appsignal.com/ruby/integrations/delayed-job.html#procs-as-jobs-with-display_name )
we recommend to define an `appsignal_name` method on a Struct to get a proper
job name for Delayed Job.

However our `extract_value` method checks if the given argument is a hash first
and attempts to read the value from that if it responds to `:[]`.

Since a Struct responds to `:[]` the `appsignal_name` method is never called.

This commit updates the `extract_value` method to first check if the given
arg responds to `:[]` and if it can't find the value in the Hash,
it then attempts to read the value from the object.