Add default Sidekiq metrics
Track more Sidekiq specific metrics by default and send them as custom
metrics. This makes it easy to set up a new Sidekiq metrics dashboard on
AppSignal.com to track more fine grained how every worker is performing.

Some Sidekiq stats are absolute counters, values that only go up over
time and are not about what has been processed in the last timeframe. To
send data about what has been processed since the minutely probe last
checked, keep track of what the previous probe call recorded and
subtract that from the new value. That way we get the delta (difference)
of those two values and can send that to AppSignal.

The cache is stored with a separately given cache_key to the
`gauge_delta` helper method, because creating our own based on the
metric key name and tags is much slower. It has to convert a Hash to
something that can be used as a cache key, which will always take
longer, and will slow down even more the more tags are used.

The metrics send as deltas from themselves:

- processed
- failed
- dead

Some keys are also not named 1 to 1 copies of the Sidekiq keys. First we
make all our keys singular. Then some other changes:

- dead -> died (jobs died since the last probe)
- retry -> retry_queue (they are not currently being retried, but are
  queued to be retried)