Add wait_for helper that checks conditions in spec
Only using `sleep X` is a little brittle as some machines are slower
than others. On faster machines you don't want to wait for the longest
time it can possibly take. So instead perform more checks in the helper
and set a maximum time after which it should have been completed.

The `wait_for` helper will check a condition for 5 seconds before giving
up and raising an error, causing the spec to fail.