Ensure only one Minutely thread can run at a time
In the test suite, whenever `Minutely.start` is called a new thread is
started that loops all the probes.

While this usually doesn't happen in an app, in the test suite a
minutely thread is started for every test that calls `Minutely.start`.
This will become worse when we enable minutely probes by default, one
will start for all specs that call `Appsignal.start`, which are a lot
more specs than call `Minutely.start` now.