Change probe registration method to use names
Register a minutely probe with a name so it's easier identifiable and
can be overwritten by another probe.

```
probe = lambda { Appsignal.set_gauge "my_gauge", 10 }

\# This is deprecated now
Appsignal::Minutely.probes << probe

\# Please use this new method of registering probes
Appsignal::Minutely.probes.register :my_probe, probe
```

This allows users to overwrite any of our probes with a new instance
that has custom config for the probe, in case the default config does
not work for their scenario.

This will allow users to quickly get the probe to work and not have to
rely on us to ship a new gem before they can test a certain probe.