Allow classes to be registered as probes
By only registering classes we can postpone the initialization of a
probe until it's needed.

For example, when `enable_minutely_probes` is `false` it won't be
initialize and we don't need to check the config in a hook whether or
not to require certain dependency files, etc.

Users can still register instances or lambdas. This option just allows
us to lazy initialize probes so that users can overwrite probes if the
default probe requires some config.

```
\# We will register it as such:
Appsignal::Minutely.probes.register :sidekiq,
  Appsignal::Hooks::SidekiqProbe

\# And this allows users to overwrite the default (future) Sidekiq
\# hook we would ship uninitialized with the gem with user
\# specific config
Appsignal::Minutely.probes.register :sidekiq,
  Appsignal::Hooks::SidekiqProbe.new(:hostname => "10.0.0.1")
```