Add hostname as tag to SidekiqProbe metrics (#471)
Multiple instances of Sidekiq using their own Redis databases can report
for the same app. To give the user meaningful metrics, that aren't
overwritten by the last instance that reports for that minute, we need
to differentiate between the hosts for which the metric applies. Which
means adding the hostname as a tag to the metric.

We try to ask the Sidekiq Redis client connection which host it's
connected to. In case this doesn't work we'll allow the user to
override the hostname config for the probe. (This is not supported in
this commit, because users can't initialize the probe (yet).)