Disable minutely probes by default in specs
On JRuby the timings seem different which is why tests are failing with:

```
RuntimeError: can't add a new key into hash during iteration
```

while initializing the probes while no probe should be running.

This is probably because the probes are now enabled by default and not
all tests stop the probe. This change ensures that no probe is continued
after a test and none is active before another test runs.

By disabling the minutely probes in the tests by default, it makes it
easier to have a reliably spec suite without every spec having to wait
for a minutely probe thread that most likely wasn't even used in the
spec.

Speed up the test suit (and make it a bit more stable on JRuby) by
disabling it by default in the test suite, while it's active in the
normal gem environment.

We just start and stop AppSignal itself way, way more often than in any
normal environment.