Fix Transaction.pause! (#482)
`Appsignal.without_instrumentation` was using `Transaction.pause!`,
which no longer worked. Events were still registered on the transaction
even though the code was wrapped in `Appsignal.without_instrumentation`
and the transaction was paused.

This change adds a guard for every Transaction method that does
something with events.

Add a spec that tests if the events are actually registered on the
transaction or not, using `Transaction#to_h`.