Fix Puma.stats calls
On the first call this would result in the following error:

```
Error in minutely probe 'puma': undefined method `stats' for nil:NilClass
```

This is because we want to start collecting metrics before Puma's
launcher has fully initialized the server yet.

Ignore the first call if it raises an internal error on a `nil` object,
the missing `Puma.@stats_object`, and try again next probe tick.