Rename PLATFORM constant for JRuby installation
On JRuby there is already a `PLATFORM` constant, and it prints the
following warning on (manual) installation.

```
rake extension:install
ext/base.rb:13: warning: already initialized constant PLATFORM
```

Rename the constant to `AGENT_PLATFORM` so we don't overwrite the
`PLATFORM` on JRuby and fix the warning printed on (manual)
installation.