Wait with the initial minutely probe run
When AppSignal is started the Minutely probe immediately fires, without
a delay. This may cause issues for when the library we're fetching data
from does not have a connection set up or initialized the class we're
asking data from.

To avoid logging errors on the first run, wait with the first run until
it's actually the start of a minute. If the app is in the last half of a
minute, wait another additional minute until running the probes.