Fix brittle method call assertions in threads (#513)
Apparently `ordered` doesn't really work well with interleaved method
calls, so don't assert on that.

And rather than expecting the method call in the before block for all
specs, only assert the actual calls in the specs where we're testing it.

Using spies like this seems more reliably; we should have less chance
that this spec fails now, based on my local testing. I was also only
able to reproduce this more frequently with gemfiles such as
"webmachine", not sure why.