Remove class variables from Minutely probes class
We don't need class variables for the Minutely probes class and can
instead use class instance variables.

This means that sub classes from the `Minutely` class, of which there
are none, will no longer inherit the class variables. Instead the
`thread` and `probes` variables will only be available directly on the
`Minutely` class. Since there is only one (instance) of the class (the
class itself) there will be difference in usage.