Remove class variables from EventFormatter class
We don't need class variables for the EventFormatter class and can
instead use class instance variables.

This means that sub classes from the `EventFormatter` class, which are
deprecated, will no longer inherit the class variables.

To make sure the deprecated method still works, for now, any child
class of `EventFormatter` will forward the call to the `EventFormatter`
class to make sure the event formatters list is only updated on the
`EventFormatter` class.