Don't raise error on log file parsing error (#517)
When there's an error in the `appsignal.yml` file, do not raise an error
up to the app, but instead log the error and print it to STDERR.

This prevents the AppSignal config initialization from crashing the app,
so it still runs, just without the AppSignal config from the
`appsignal.yml` file. The message printed to STDERR should be enough to
indicate to the user there was a problem. We also log it to the
`appsignal.log` file so it shows up in the diagnose report.

By rescuing the error the diagnose command will also be able to finish
in this scenario allowing us to debug other issues, without having to
fix the config file first.