Rescue no user/group error
Return `nil` rather than let the error bubble up.

When a user or group cannot be found it's usually a mounted volume from
a host system where the guest system does not know the user/group that
own the files/directories.

We can probably store this in a better way, e.g. a hash containing the
error and the arguments given, but this should be consistently set for
all values, not only this one. So we look at that in a future release.