Fix error in AS when transaction get completed (#532)
If an AppSignal transaction is closed in an instrumentation block (which
shouldn't happen), the transaction can no longer be called to finish an
event.

```
ActiveSupport::Notifications.instrument("foo") do
  Appsignal::Transaction.complete_current!
end
```

To fix the error that would occur, fetch the transaction again from
`Appsignal::Transaction.current` instead of using the `transaction`
variable that was available. If the transaction was closed inside the
instrumentation block, it will return a `NilTransaction` which will
ignore the method call instead of raising an error.