Fix error when transactions are completed
If an AppSignal transaction is closed in an app using the
GenericInstrumentation middleware AppSignal would raise an error.

To fix the error that would occur, fetch the transaction again from
`Appsignal::Transaction.current` instead of using the `transaction`
variable that was available. If the transaction was closed inside the
instrumentation block, it will return a `NilTransaction` which will
ignore the method call instead of raising an error.