Add AppSignal transaction test helpers
When an AppSignal transaction gets completed it is no longer available
in the extension for it to return a JSON representation of it.

The Ruby Transaction class will get `null` back from the extension which
the `JSON` library can't parse, which raises an error.

This makes it difficult in specs to use the
`Appsignal::Transaction#to_h` helper to get a Ruby Hash based on the
extension transaction JSON. Specs that use the
`Appsignal::Transaction#to_h` helper previously stubbed the
`Appsignal::Extension::Transaction#complete` method call so it is
possible to fetch the JSON during the spec.

This is a cumbersome process which makes it difficult to use the `#to_h`
in specs. Something which we prefer to do as expecting method call is a
fragile assertion, especially when `and_call_original` is not used on
the expectation.

Instead of stubbing `Appsignal::Extension::Transaction` call on the
`finish` and `complete` methods override them and add some extra
behavior in the test environment. This extra behavior requests and
stores the transaction JSON representation before the `complete` call is
made to the extension. This means we have access to the JSON while also
completing the transaction in the extension.

Additionally it doesn't break the
`Appsignal::Extension::Transaction#finish` call, which is the case with
using the RSpec method call expectation. So it would still send all the
arguments of the `Appsignal::Extension::Transaction#finish` method to
the extension.

This can now be used with the `keep_transactions` spec helper. See the
diff of this commit for some examples.

This new testing helper is not active used by default for all specs as
it may have a performance impact when it applies to all transactions in
the test suite. Not all tests rely this behavior, so only use it in
those that need it.