Fix Object integration specs
When a transaction is created before the config is set (even in a spec)
it fails on a missing configuration option, the
`enable_gc_instrumentation` option being requested in
`Appsignal::Transaction.garbage_collection_profiler`.

This spec doesn't fail when run in the entire test suite, because other
specs are run beforehand that set the `Appsignal.config` with *any*
value. It does fail when run on its own.

To make our test suite more resilient and allow it to run in a random
order in the future, fix the setup in this spec file.