Assert if transactions were completed
Test if a transaction was actually completed in an integration. This
allows us to test if we're still starting and finishing whole
transactions. This was previously asserted by listening to the
`Appsignal::Transaction#complete` method call, which is fragile without
at least and `and_call_original` call.

This change tracks whether or not a transaction was completed by
listening for the `Appsignal::Extension::Transaction#complete` call and
setting an instance variable. The original logic remains intact.

If it was completed can be asked by calling
`Appsignal::Extension::Transaction#_complete?`.

Use the `be_completed` matcher on a transaction to assert if it was
completed or not. Hide the internal classes search in the matcher so we
don't have to repeat it in every spec. Updated specs that are fully over
to the new test helpers.