Refactor Appsignal::Testing helpers
Use modules and include (prepend/extend) existing classes rather than
reopening classes and monkeypatching methods (aliasing the old method
and creating a new method that wraps around it.

With modules we can more explicitly load in the helper logic and
methods. It also allows us to do away with the aliasing of methods and
rather call `super` inside the methods we're enhancing with testing
logic.