Fix Rakefile loading on Ruby 1.9.3
When loaded on Ruby 1.9.3 the Rakefile errors while defining the
`biuld:<lang>:gem` tasks. Not sure why, it complains about something
being a Hash while it should be a String. We don't define any hashes, so
it must be something internal.

Since we're not planning on building and publishing the gem from Ruby
1.9.3, ignore the error so the Rakefile can be loaded.
To not completely ignore it, print a warning message that an error
occurred and the tasks won't be available.

This at least allows us to run `rake extension:install` and test against
Ruby 1.9.3.