Improve Puma compatibility
Adding a `before_fork` proc that starts our Minutely Puma probe.

When `daemonize true` is set in the Puma config, this ensures that
a new minutely probe is started after daemonizing the process.

The probe needs to live in the "master" process,
to gather all the wanted metrics.

This doesn't cover all cases, just the one(s) where Puma is started
from it's own CLI.