Mark helper modules as not private
Previously marked as private incorrectly to discourage the usage of
the modules directly. This was not meant to also mark them as private
through the usage of the `Appsignal` module, which is the recommended
approach.

Yard however, does not show them directly on the `Appsignal` module's
documentation page, but links to the helper modules instead, which
showed the private method warning.