Check set_error arguments for Exceptions
It was possible to pass in a non-Exception object to
`Appsignal.set_error` and `Appsignal::Transaction#set_error`.

This would then error on `object.backtrace` if the object didn't
implement a backtrace method.

Add additional checks and logging to these methods when the argument
isn't a valid Exception object instance. It's not possible to pass in
Exception classes either, these have the same problems as normal Ruby
objects as arguments.

Update `Appsignal.send_error` which did have these checks to log the
same kind of error message.