Handle missing application.rb for Rails installer
When an app is detected as a Rails app by the presence of the `rails`
gem we load the `config/application.rb` file, assuming it exists. Catch
the scenario where it doesn't.

It doesn't error anymore, but doesn't show the desired output perhaps,
it only ends the installer without prompting the user for a custom app
name. This should be picked up as part of #315.

Co-authored-by: Mark <birmanmark@gmail.com>