Handle Rails application.rb load errors in install
When the Rails' application.rb references code that isn't available to
the AppSignal installer, such as loading another gem, it results in a
`StandardError` subclass. When this occurs, or the file loading fails in
general with `LoadError` rescue the installation and prompt the user for
an app name.

I've changed the `require` call to a `load` call for the
`config/application.rb` file, because in our tests this file gets
required multiple times. The second time it's being required it won't be
loaded again, because it's already required.

To ensure the file is (attempted) to be loaded, use Ruby's `load`
instead. This will load the file again even if it's loaded already. This
will make our tests more accurate and shouldn't impact the installer
since the file is only loaded once in the installer.