Check for empty Push API key in config
Don't make AppSignal active if the key is empty or only contains spaces.
This should make the active check more reliably as empty values are not considered valid.

For example:

```yaml
push_api_key: "<%= ENV['APPSIGNAL_PUSH_API_KEY'] %>"
```

This in the `config/appsignal.yml` file returns an empty string if the
Push API key is not set. But we did detect it as a valid value.