Add Puma plugin to start probe in main process
Add a new Puma plugin to start the Puma probe only in the Puma's main
process. The old solution started the Puma probe in every worker.
(Duplicate stats from X workers weren't an issue.) Puma only exposed the
worker stats in the main process, and the workers all reported `0` as
values.

As a fallback we detect if users have loaded the plugin with the
{APPSIGNAL_PUMA_PLUGIN_LOADED} constant. If it hasn't it will fall back
on the old behavior which works for non-cluster mode setups.

We prefer people use the AppSignal Puma plugin from now on. The fallback
on the old behavior is only there when users relied on our *magic*
integration. The minutely probe thread will still run in Puma workers,
for other non-Puma probes, but the Puma probe only runs in the Puma main
process.

To add the AppSignal plugin to Puma add this to your `config/puma.rb`
file (or other file location if you have configured the config file
location).

```
plugin :appsignal
```

Co-authored-by: Robert Beekman <robert@matsimitsu.nl>