Remove SSL version lock in transmitter
By removing the version lock we will no longer be limited to TLSv1, this
has known security issues.

By not specifying it, it will work with TLSv1.2 as well.

The `ssl_version` option itself is also deprecated. Instead we could use
the `min_option` version, but we think the Ruby default should be good
enough and more future proof.