Fix wait_for helper (#581)
It would hide the error if the error would occur for 5000 retries.

It always iterated once more than the max retry count. Fix that, but
also make the assertion for showing the error also show the error when
it makes more iterations in the future, so that this doesn't happen
again.