Fix exception handling of config file issues
When the `config/appsignal.yml` file would raise an error we print the
error name and error message.

Unfortunately do to the chosen test error it was not caught that
`Exception#name` is not an available method to ask for the Exception
class name. This would cause an error for other errors that were not an
`NameError`.
`NameError` does listen to the `#name` method, returning the constant
name that was missing in the test.

To fix this, update the class name call to the Exception class name,
instead of the Exception instance.