Add logger that can warn once, then use debug. (#585)
* Add logger that can warn once, then use debug.

Right now we use `logger.warn` for a number of messages that occur
very often in client apps.

To prevent a huge amount of log lines, this subclassed logger as a new
method called `warn_once_then_debug` to only "warn" once, then default to "debug".

* Fix tests and linting violations

* Update test to not stub method calls

Assert the output, not method calls. When asserting method calls we
can't be sure if it still works in the future, even if the methods are
called.

This way we assert if the log contents actually match the expected
values.

* Clean up appsignal spec logger assertions

No need for a separate test case for the logger assertion. It's a base
requirement we can add to the test setup block. Saves rerunning all the
setup for those extra test cases.

Co-authored-by: Tom de Bruijn <tom@tomdebruijn.com>