Move test build to Semaphore (#587)
* Move test build to Semaphore

Use Semaphore instead of TravisCI. The Semaphore build seems a lot more
stable than the TravisCI build. On TravisCI we get random failures that
require manual restart of individual jobs. We have to watch the build
and retries to make sure the build is green.
The overal build on Semaphore is also done quicker.

I've split the build up per Ruby version. This is because Semaphore
supports a maximum of 50 jobs per task. Every Ruby version first checks
the build for the `no_dependencies.gemfile`. This tests if the build at
all works for that Ruby version. If it doesn't it will not even build
the gem builds for that Ruby version, finishing the build quicker.

Bonus: We can run the test suite on Ruby 1.9.3, while we still sort of
support it. No more accidental breaking the build again.

* Update build badge to Semaphore [ci skip]