Add grape.skip_appsignal_error request env
Listen to `grape.skip_appsignal_error` request env to ignore errors if
it's set to `true`. This way users can ignore errors in their
app for specific occurrences, not just ignore all occurrences with
the `ignore_errors` option.

This solution is the same as is available in Sinatra with
`sinatra.skip_appsignal_error`.