Print job count output in semaphore.yml generate
This way we can more easily tell if we see a big increase/decrease in
the number of jobs after changing the `build_matrix.yml` file. This
restores the previous behavior we had for Travis, but because of
Semaphore's block structure we have to count the jobs per task first.

[ci skip]