Use Module.prepend for net_http on newer Rubies (#597)
In combination with the `http_logger` gem the AppSignal integration
causes a `SystemStackError (stack level too deep)` error.

This is because the `http_logger` gem uses `Module.prepend` and the
AppSignal gem uses aliases to monkeypatch the Net::HTTP behavior.

By also switching to `Module.prepend`, which we should probably use
moving forward, this problem can be avoided.

The only issue is that `Module.prepend` is not available on Ruby 1.9, so
we need to make an exception for that while we support it. On Ruby 1.9
it will continue to use the alias approach to integrate AppSignal.