Track queue time regardless of namespace (#602)
* Track queue time regardless of namespace

We had a check that only kept track the queue time for the web and
background namespaces. Remove that check now that we allow users to use
custom namespaces.

Prefer the web namespace HTTP header method of setting the queue time
(for things like admin namespace) and fall back on the background
namespace method (setting the `:queue_start` value on the transaction
env).

Depending on the method with which the value is set a different
calculation to set the time in milliseconds is done.

- HTTP header: expected value is in either milliseconds or microseconds.
- `:queue_start` key value on the transaction env: expected value in in
  seconds.

Update specs to always return the same magnitude of value for every
test case. Some tests didn't call `.to_i` on the value set on the
HTTP header, which caused the value to be read a factor of 100 lower
than the test cases for the `:queue_start` env key value. Now all test
cases are using the same magnitude of value for the queue times. Also
commented which calculations are done one which lines to hopefully make
this more understandable in the future.

* Document set_http_or_background_queue_start

[ci skip]