Fix issue where Delayed Job job objects responding to Hash accessor assumes string return value
With our Delayed Job integration it's possible to override the job's class/method name by implementing
an `appsignal_name` method or setting the value in the Job Hash.

Our `extract_value` code first attempts to read the value from the Hash,
and then attempts to read it as an object accessor.

If objects respond to `[]` with something other than a String, this could crash the plugin.

This commit adds more checking of return values from `extract_value` and only attempts to set the
class/method if the return value is a String. Otherwise it will return the given default value,
or `unknown` if no valid default value is found.