Revert DelayedJob class/method name extraction method to generate the same results as before this fix.
Code downstream relies on the fact that class methods (`Foo.bar`) are emitted as `["Foo.bar", nil]` for class/method name.
This commit ensures that this behavior is re-implemented in the DelayedJob class/method name extraction logic to ensure that
we don't suddenly generate a ton of new incidents with different "action" names.