Fix ResqueActiveJobPlugin specs
Remove Resque presence check, it's not present in the ActiveJob specs
for it. The DependencyHelper would check for a scenario that never
occurred.

This meant the tests were never run in either the Resque gemfile or
Rails gemfiles setups.

Since we don't need Resque in this spec, since we stub it out, remove
the DependencyHelper check.