Add helpers for environment metadata collection
Add extension function and helper method to send environment metadata
from the Ruby gem to the extension.

This data will be use to improve AppSignal by tracking what AppSignal
integrations, and versions of these integrations, are actually used.

This is the basic implementation of the functionality. Documentation
about what data is collected and what it is used for will follow later.