Add more defensive checks on report metadata
Make sure that the keys and values that are send to the extension
function are of the String type. If another type of value is given it
can cause a segfault, crashing the entire application. Add logging for
every step that the value can be something else than a String.

Also capture and log errors that may occur during this process.