Report which AppSignal supported gems are present
Collect metadata (name and version) about gems AppSignal integrates
with.

Ask Bundler about the gems that are available in the app's "bundle", and
then loop over the SUPPORTED_GEMS constant to only report about the gems
AppSignal integrates with. This way we do not collect data about _all_
apps in an app's bundle, which can contain things like private gems we
do not want to report.

I decided to use the data available in Bundler, rather than asking every
gem separately by calling the value of a `VERSION` constant. There are a
lot of different ways these constants are defined (e.g. `VERSION`,
`VERSION_STRING`) and I even found a gem that did not have one at all.
Using this method we can be sure that all the data is reported in the
same way, and there's less change of the reporting breaking when a gem
updates how they define their `VERSION` constant.

Another problem with collecting the data in the integrations themselves
is that, some integrations for gems are not installed automatically.
The Capistrano integration for example only loads when `require`d by an
app's `Capfile`. And then only when that `Capfile` itself is loaded,
such as for a deploy. This would make the reporting for these kinds of
integrations less accurate because the time we store this data may be
shorter than it is called. Creating inaccurate reporting.

An alternative approach, or fallback, would be to ask Ruby which gems
are loaded at load time, but that has the same problem as the method
that only sets the metadata on load of an integration, so it was not
chosen.

Report the actioncable and activejob gems separately as they can be used
outside of Rails and still have the AppSignal instrumentation installed.