Report which integrations are enabled
Some AppSignal integrations are not installed automatically, or can be
manually disabled with configuration options, e.g.
`instrument_net_http`. Report the usage of these integrations separately
using the `report_enabled` helper method.

This way we can make the distinction between an AppSignal supported gem
being present in an app's "bundle" and the integration actively using
that integration.

Using this method we can also report metadata about AppSignal
integrations that do not rely on a gem, such as the Object method
instrumentation integration.