Add send_environment_metadata option
Allow users to disable environment metadata transmission in the agent
with the `send_environment_metadata` config option. This config option
is read by the Ruby gem, which sets it in the app "ENV" that is read by
the extension and agent as configuration source.

By default the environment metadata transmission is on, but this gives
apps more control over what environment metadata they share.

The metadata is still sent from the Ruby gem to the extension and agent
using the `appsignal_set_environment_metadata` function in the
extension, but this data won't be transmitted to AppSignal servers.

The handling of this config option is added in the agent, as the agent
also collects some metadata. If we would only implement this config
option in the Ruby gem the agent environment metadata would still be
sent. This ensures the handling for this config option is in one place,
the agent.