Update failing JSExceptionCatcher specs
Update the excepted deprecation message to the new message.

Allow the deprecation message to contain the phrase "error" by
allowlisting the `enable_frontend_error_catching` config option as an
accepted phrase in the `silence` helper method. The `silence` helper
raises an error if it finds the phrase "error", which failed most specs.