Add ActiveJob integration for all adapters
The ActiveJob integration will add support for ActiveJob adapters that
do not have ActiveJob support themselves in AppSignal integrations.
Other adapters are listed here:
https://api.rubyonrails.org/classes/ActiveJob/QueueAdapters.html

We will now support these library when used through ActiveJob.

Sidekiq, Delayed Job, Resque all have their own ActiveJob support right
now. In the future we can take out the library specific ActiveJob
support, they should no longer be necessary. In some cases we will still
need to catch when ActiveJob values are found, in order to not error
while instrumenting the jobs.

The ActiveJob integration runs within the background job libraries. That
means that if another integration, like the one for Sidekiq is wrapping
the ActiveJob job, we need to check if a transaction is already active
or not. If a transaction is active, reuse that transaction and add more
data to it, but do not create a new one or close the active transaction.

If we would create a new transaction AppSignal would log a warning and
reuse the same transaction anyway, but if the ActiveJob integration
would "complete" the transaction, any event or data set by the Sidekiq
integration would no longer be recorded.