Reduce number of duplicate specs
The shared examples called `sidekiq metadata` were being run in the
same type of context. The "with an error" and "without an error"
contexts did not meaning fully affect the tests being run, there are no
`before` or `after` blocks that affect the tests being run. They were
identical test runs which we can avoid by moving them out of the shared
example group.

Decreases the number of specs from 57 to 35. And the run time from
`0.31` down to `0.16`.