Update Sidekiq ActiveJob integration
Remove most ActiveJob logic from the AppSignal Sidekiq middleware. That
is now handled in the separate ActiveJob integration. Only the arguments
unwrap have a special ActiveJob case so that the arguments set from the
ActiveJob integration aren't overwritten by the Sidekiq integration.

This is not an issue for the action name, that's only set if the action
name is `nil` at that point. So when the action name in the Sidekiq
integration returns
`ActiveJob::QueueAdapters::SidekiqAdapter::JobWrapper` does not
overwrite the action name set in the ActiveJob integration.

Load Sidekiq in the test suite itself, because I can't mimick the
middleware in a test that only interfaces with the AppSignal
SidekiqPlugin middleware. Using the Sidekiq test helpers we can test the
jobs without sending them to Redis and processing them async, making the
tests run fast.

The AppSignal middleware does need to be included manually in this test
environment, because Sidekiq doesn't use the normally configured
middleware by default. We test somewhere else if the AppSignal
middleware is configured using the installer.

For the tests, update the "sidekiq/api" stub to more closely match the
actual file in Sidekiq. Define the `Sidekiq::Stats` class instead of
`Sidekiq::API`, which doesn't exist, and we actually rely on
`Sidekiq::Stats` being defined in the probe.