Add support for ActionMailer::MailDeliveryJob
The `ActionMailer::MailDeliveryJob` new in Rails 6 will be the default
in Rails 6.1.

Add support for it and add more testing for the different argument types
(method or parameterized).

Decided not to "clean up" those arguments before storing it on the
transaction, we can always decide to do that later, and it avoids
adding complexity now that could break things.