Only restore CI caches on exact match (#626)
When something in the cache key has changed, do not use an older cache.
This can cause problem with gem updates not being picked up.

When no cache is found the install process shouldn't slow down the build
a ton that we have to wait forever for it to complete. So we can take
the extra time there for stability overall.