Add support for queue time to ActiveJob
Use the `Transaction#set_queue_start` method to set the queue time
directly instead of through
`Transaction#set_http_or_background_queue_start`, because it won't get
set on the transaction environment all the times. A transaction can
already be created for the ActiveJob because of a wrapper integration
such as Sidekiq, DelayedJob and Resque. Then if it's only set on a new
transaction, it won't work for those integrations.

Document the previously undocumented Transaction#set_queue_start method
and update its tests code style.